requirements.txt
fastapi>=0.116.0
uvicorn>=0.35.0
openai>=1.108.0
langgraph>=0.6.0
langchain-core>=0.3.0
pandas>=2.3.0
openpyxl>=3.1.5
pydantic>=2.11.0
python-dotenv>=1.1.0
requests>=2.32.0
pymupdf>=1.26.0
beautifulsoup4>=4.12.0